Split Results into a container and a component
