+++ js/src/components/contentAnalysis/ReadabilityAnalysis.js
@@ -9,7 +9,7 @@ import { isNil, get } from "lodash-es";
 
 /* Internal components */
 import ScoreIconPortal from "../portals/ScoreIconPortal";
-import Results from "./Results";
+import Results from "./containers/Results";
 import Collapsible from "../SidebarCollapsible";
 import getIndicatorForScore from "../../analysis/getIndicatorForScore";
 import { getIconForScore } from "./mapResults";
