+++ js/src/components/contentAnalysis/containers/Results.js
@@ -1,0 +1,26 @@
+import { withSelect, withDispatch } from "@wordpress/data";
+import { compose } from "@wordpress/compose";
+import Results from "../Results";
+
+export default compose( [
+	withSelect( select => {
+		const {
+			getActiveMarker,
+		} = select( "yoast-seo/editor" );
+
+		return {
+			activeMarker: getActiveMarker(),
+		};
+	} ),
+	withDispatch( dispatch => {
+		const {
+			setActiveMarker,
+			setMarkerPauseStatus,
+		} = dispatch( "yoast-seo/editor" );
+
+		return {
+			setActiveMarker,
+			setMarkerPauseStatus,
+		};
+	} ),
+] )( Results );
